refactor(#30): prefetchQuery 함수 위치 변경
